spec: configre --enable-readline --with-readline-lib=-lreadline
New configure script thinks that, along with -lreadline, we also
need to link with the library which provdes tgetent(), which
cofigure thinks must be ncurses.  However, tegetent() is provided
by libtinfo.  To drop all this smart logic, I have to use configure
option.  The result is no build dependency on libncurses-devel.